Feat: Adds links drawer opening to navigation icon on top-app-bar
previously when this was attempted, the id of the top-app-bar header was missing, and that is why the javascript did not function.
Everythin seems to function like it should now.  However, there is still a css bug that is causing a vertical scroll bar to appear.
And the top-app-bar seems to have a left padding or margin that needs to be fixed.