Feat: creates modal drawer for mobile first design
Since I will be designing the app for mobile first,
and the material design guidelines are to use mdc-drawer--modal for mobile
designs web designs.  I wanted this to be where I start.
I will need to add some css breakpoints and change the class of the drawer
depending on the screen-size.