Feat: updates to webpack 4, includes modular sass partials, modularizes webpack configs
It is not ideal, but I started hacking away and now I am making a large commit that involve many
logical changes in one commit.  In this commit, when running npm run build, a dist folder will be
created with and index.html, bundle.css, and app.bundle.js. this index.html file will load
in a browser and work like it should.  However, I have not tested the development environment yet.