Ccreates dictionary of input field data for dry error reporting
I want to be able to iterate through the values and then provide the proper
error reporting.  I think using a dictionary could be helpful because of the get
funciton that allows a default to be used which might save a few lines when there
are null's in the dict values.